Refactor Processor and Aggregator
