+++ kmock-processor/src/test/kotlin/tech/antibytes/kmock/processor/KMockAggregatorCommonSpec.kt
@@ -1,0 +1,518 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache v2.0
+ */
+
+package tech.antibytes.kmock.processor
+
+import com.google.devtools.ksp.processing.KSPLogger
+import com.google.devtools.ksp.processing.Resolver
+import com.google.devtools.ksp.symbol.ClassKind
+import com.google.devtools.ksp.symbol.KSAnnotated
+import com.google.devtools.ksp.symbol.KSAnnotation
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSFile
+import com.google.devtools.ksp.symbol.KSType
+import com.google.devtools.ksp.symbol.KSTypeAlias
+import com.google.devtools.ksp.symbol.KSTypeReference
+import com.google.devtools.ksp.symbol.KSValueArgument
+import io.mockk.Runs
+import io.mockk.every
+import io.mockk.just
+import io.mockk.mockk
+import io.mockk.verify
+import org.junit.jupiter.api.Test
+import tech.antibytes.kmock.MockCommon
+import tech.antibytes.kmock.fixture.StringAlphaGenerator
+import tech.antibytes.util.test.fixture.fixture
+import tech.antibytes.util.test.fixture.kotlinFixture
+import tech.antibytes.util.test.fixture.qualifier.named
+import tech.antibytes.util.test.fulfils
+import tech.antibytes.util.test.mustBe
+
+class KMockAggregatorCommonSpec {
+    private val fixture = kotlinFixture { configuration ->
+        configuration.addGenerator(
+            String::class,
+            StringAlphaGenerator,
+            named("stringAlpha")
+        )
+    }
+
+    @Test
+    fun `It fulfils Aggregator`() {
+        KMockAggregator(
+            mockk(),
+            mockk(),
+            mockk(),
+            emptyMap(),
+            emptyMap(),
+        ) fulfils ProcessorContract.Aggregator::class
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it resolves the Annotated Thing as ill, if no KMockAnnotation was found`() {
+        // Given
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+
+        val annotation1: KSAnnotation = mockk()
+        val annotation2: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation1)
+            yield(annotation2)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every {
+            annotation1.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns fixture.fixture()
+
+        every {
+            annotation2.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns fixture.fixture()
+
+        every { source.annotations } returns sourceAnnotations
+
+        // When
+        val (illegal, _, _) = KMockAggregator(
+            mockk(),
+            mockk(),
+            mockk(),
+            emptyMap(),
+            emptyMap(),
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        illegal mustBe listOf(source)
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it filters all ill Stub Annotations`() {
+        // Given
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+
+        val annotation: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every { annotation.arguments } returns emptyList()
+
+        every {
+            annotation.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns MockCommon::class.qualifiedName!!
+
+        every { source.annotations } returns sourceAnnotations
+
+        // When
+        val (illegal, _, _) = KMockAggregator(
+            mockk(),
+            mockk(),
+            mockk(),
+            emptyMap(),
+            emptyMap(),
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        illegal mustBe listOf(source)
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it filters all non class types and reports an error`() {
+        // Given
+        val logger: KSPLogger = mockk()
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+        val file: KSFile = mockk()
+
+        val annotation: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        val type: KSType = mockk(relaxed = true)
+        val declaration: KSTypeAlias = mockk(relaxed = true)
+        val arguments: List<KSValueArgument> = mockk()
+
+        val values: List<KSType> = listOf(type)
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every {
+            annotation.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns MockCommon::class.qualifiedName!!
+
+        every { source.annotations } returns sourceAnnotations
+
+        every { annotation.arguments } returns arguments
+        every { arguments.size } returns 1
+        every { arguments.isEmpty() } returns false
+        every { arguments[0].value } returns values
+        every { type.declaration } returns declaration
+        every { file.parent } returns null
+        every { source.parent } returns file
+
+        every { logger.error(any()) } just Runs
+
+        // When
+        KMockAggregator(
+            logger,
+            mockk(),
+            mockk(),
+            emptyMap(),
+            emptyMap(),
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        verify(exactly = 1) { logger.error("Cannot stub non interfaces.") }
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it filters all implementation class types and reports an error`() {
+        // Given
+        val logger: KSPLogger = mockk()
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+        val file: KSFile = mockk()
+
+        val selection = listOf(
+            ClassKind.CLASS,
+            ClassKind.ENUM_CLASS,
+            ClassKind.ENUM_ENTRY,
+            ClassKind.OBJECT,
+            ClassKind.ANNOTATION_CLASS
+        )
+
+        val selector = fixture.random.access { it.nextInt(0, selection.lastIndex) }
+
+        val annotation: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        val type: KSType = mockk(relaxed = true)
+        val declaration: KSClassDeclaration = mockk(relaxed = true)
+        val arguments: List<KSValueArgument> = mockk()
+
+        val values: List<KSType> = listOf(type)
+
+        val className: String = fixture.fixture(named("stringAlpha"))
+        val packageName: String = fixture.fixture(named("stringAlpha"))
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every {
+            annotation.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns MockCommon::class.qualifiedName!!
+
+        every { source.annotations } returns sourceAnnotations
+
+        every { annotation.arguments } returns arguments
+        every { arguments.size } returns 1
+        every { arguments.isEmpty() } returns false
+        every { arguments[0].value } returns values
+        every { type.declaration } returns declaration
+        every { declaration.classKind } returns selection[selector]
+
+        every { declaration.parentDeclaration } returns null
+
+        every { file.parent } returns null
+        every { source.parent } returns file
+
+        every { declaration.qualifiedName!!.asString() } returns className
+        every { declaration.packageName.asString() } returns packageName
+
+        every { logger.error(any()) } just Runs
+
+        // When
+        KMockAggregator(
+            logger,
+            mockk(),
+            mockk(),
+            emptyMap(),
+            emptyMap(),
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        verify(exactly = 1) { logger.error("Cannot stub non interface $packageName.$className.") }
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it returns all found interfaces`() {
+        // Given
+        val logger: KSPLogger = mockk()
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+        val file: KSFile = mockk()
+
+        val annotation: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        val type: KSType = mockk(relaxed = true)
+        val declaration: KSClassDeclaration = mockk(relaxed = true)
+        val arguments: List<KSValueArgument> = mockk()
+
+        val values: List<KSType> = listOf(type)
+
+        val className: String = fixture.fixture(named("stringAlpha"))
+        val packageName: String = fixture.fixture(named("stringAlpha"))
+
+        val genericResolver: ProcessorContract.GenericResolver = mockk()
+        val generics: Map<String, List<KSTypeReference>>? = if (fixture.fixture()) {
+            emptyMap()
+        } else {
+            null
+        }
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every {
+            annotation.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns MockCommon::class.qualifiedName!!
+
+        every { source.annotations } returns sourceAnnotations
+
+        every { annotation.arguments } returns arguments
+        every { arguments.size } returns 1
+        every { arguments.isEmpty() } returns false
+        every { arguments[0].value } returns values
+        every { type.declaration } returns declaration
+        every { declaration.classKind } returns ClassKind.INTERFACE
+
+        every { declaration.parentDeclaration } returns null
+
+        every { file.parent } returns null
+        every { source.parent } returns file
+
+        every { declaration.qualifiedName!!.asString() } returns className
+        every { declaration.packageName.asString() } returns packageName
+
+        every { logger.error(any()) } just Runs
+
+        every { genericResolver.extractGenerics(any(), any()) } returns generics
+
+        // When
+        val (_, interfaces, _) = KMockAggregator(
+            logger,
+            mockk(),
+            genericResolver,
+            emptyMap(),
+            emptyMap(),
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        interfaces mustBe listOf(ProcessorContract.TemplateSource("", declaration, null, generics))
+
+        verify(exactly = 1) { genericResolver.extractGenerics(declaration, any()) }
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it returns the corresponding source files`() {
+        // Given
+        val logger: KSPLogger = mockk()
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+        val file: KSFile = mockk()
+
+        val annotation: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        val type: KSType = mockk(relaxed = true)
+        val declaration: KSClassDeclaration = mockk(relaxed = true)
+        val arguments: List<KSValueArgument> = mockk()
+
+        val values: List<KSType> = listOf(type)
+
+        val className: String = fixture.fixture(named("stringAlpha"))
+        val packageName: String = fixture.fixture(named("stringAlpha"))
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every {
+            annotation.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns MockCommon::class.qualifiedName!!
+
+        every { source.annotations } returns sourceAnnotations
+
+        every { annotation.arguments } returns arguments
+        every { arguments.size } returns 1
+        every { arguments.isEmpty() } returns false
+        every { arguments[0].value } returns values
+        every { type.declaration } returns declaration
+        every { declaration.classKind } returns ClassKind.INTERFACE
+
+        every { declaration.parentDeclaration } returns null
+
+        every { file.parent } returns null
+        every { source.parent } returns file
+
+        every { declaration.qualifiedName!!.asString() } returns className
+        every { declaration.packageName.asString() } returns packageName
+
+        every { logger.error(any()) } just Runs
+
+        // When
+        val (_, _, sourceFiles) = KMockAggregator(
+            logger,
+            mockk(),
+            mockk(relaxed = true),
+            emptyMap(),
+            emptyMap(),
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        sourceFiles mustBe listOf(file)
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+
+    @Test
+    fun `Given extractCommonInterfaces is called it returns while mapping aliases`() {
+        // Given
+        val logger: KSPLogger = mockk()
+        val source: KSAnnotated = mockk()
+        val resolver: Resolver = mockk()
+        val file: KSFile = mockk()
+        val sourceSetValidator: ProcessorContract.SourceSetValidator = mockk()
+
+        val annotation: KSAnnotation = mockk()
+        val sourceAnnotations: Sequence<KSAnnotation> = sequence {
+            yield(annotation)
+        }
+
+        val annotated: Sequence<KSAnnotated> = sequence {
+            yield(source)
+        }
+
+        val type: KSType = mockk(relaxed = true)
+        val declaration: KSClassDeclaration = mockk(relaxed = true)
+        val arguments: List<KSValueArgument> = mockk()
+
+        val values: List<KSType> = listOf(type)
+
+        val className: String = fixture.fixture(named("stringAlpha"))
+        val alias: String = fixture.fixture(named("stringAlpha"))
+        val packageName: String = fixture.fixture(named("stringAlpha"))
+
+        val genericResolver: ProcessorContract.GenericResolver = mockk()
+        val generics: Map<String, List<KSTypeReference>>? = if (fixture.fixture()) {
+            emptyMap()
+        } else {
+            null
+        }
+
+        val mapping = mapOf(className to alias)
+
+        every {
+            resolver.getSymbolsWithAnnotation(any(), any())
+        } returns annotated
+
+        every {
+            annotation.annotationType.resolve().declaration.qualifiedName!!.asString()
+        } returns MockCommon::class.qualifiedName!!
+
+        every { source.annotations } returns sourceAnnotations
+
+        every { annotation.arguments } returns arguments
+        every { arguments.isEmpty() } returns false
+
+        every { arguments.size } returns 1
+        every { arguments[0].value } returns values
+
+        every { type.declaration } returns declaration
+        every { declaration.classKind } returns ClassKind.INTERFACE
+
+        every { declaration.parentDeclaration } returns null
+
+        every { file.parent } returns null
+        every { source.parent } returns file
+
+        every { declaration.qualifiedName!!.asString() } returns className
+        every { declaration.packageName.asString() } returns packageName
+
+        every { logger.error(any()) } just Runs
+
+        every { sourceSetValidator.isValidateSourceSet(any()) } returns true
+
+        every { genericResolver.extractGenerics(any(), any()) } returns generics
+
+        // When
+        val (_, interfaces, _) = KMockAggregator(
+            logger,
+            sourceSetValidator,
+            genericResolver,
+            emptyMap(),
+            mapping,
+        ).extractCommonInterfaces(resolver)
+
+        // Then
+        interfaces mustBe listOf(ProcessorContract.TemplateSource("", declaration, alias, generics))
+
+        verify(exactly = 1) { genericResolver.extractGenerics(declaration, any()) }
+        verify(exactly = 1) {
+            resolver.getSymbolsWithAnnotation(MockCommon::class.qualifiedName!!, false)
+        }
+    }
+}
